FLASK_API/requirements.txt
#Flask==2.3.3
#Flask-RESTful==1.0.9
#Flask-Cors==3.0.10
#Flask-SQLAlchemy==3.0.5
#PyMySQL==1.1.2
#python-dotenv==1.2.2
#Werkzeug==2.3.8
#Jinja2==3.1.3
#itsdangerous==2.1.2
#click==8.2.0
fastapi==0.135.1
uvicorn==0.42.0
SQLAlchemy==2.0.48
PyMySQL==1.1.2
cryptography==46.0.5
python-dotenv==1.2.2
passlib==1.7.4
bcrypt==5.0.0
python-jose==3.5.0
pydantic==2.12.5
pydantic_core==2.41.5
anyio==4.12.1
starlette==0.52.1
typing_extensions==4.15.0
typing-inspection==0.4.2
greenlet==3.3.2
cffi==2.0.0
pycparser==3.0
rsa==4.9.1
ecdsa==0.19.1
pyasn1==0.6.3
six==1.17.0
idna==3.11

click==8.3.1
h11==0.16.0
colorama==0.4.6
annotated-types==0.7.0
annotated-doc==0.0.4